this is layout chnage
Add user name display and test comment in _Layout.cshtml

Added a `<span>` element to display the user's name dynamically
from the session variable `HttpContext.Current.Session["Name"]`.
The `<span>` is styled to be responsive, visible only on medium
or larger screens. Also added a test comment `@*this is the test*@`
for documentation or testing purposes.